+++ tests/DFE.FindInformationAcademiesTrusts.CypressTests/cypress/support/audit-page-definitions.ts
@@ -107,4 +107,11 @@ export class AuditPageDefinitions {
         this.auditHelper.auditPage('LA Maintained School Ofsted - Safeguarding and concerns', 'Schools', `/Schools/Ofsted/SafeguardingAndConcerns?urn=${laMaintainedSchoolUrn}`);
 
     }
-} 
+
+    /***
+     * Audit manage projects and cases pages
+     */
+    auditManageProjectsAndCasesPages(): void {
+        this.auditHelper.auditPageWithRole('User.Role.MPCViewer', 'Manage Projects and Cases - Overview', 'Manage Projects and Cases', '/manageprojectsandcases/overview');
+    }
+}